+++ packages/types/src/version.d.ts
@@ -1,2 +1,2 @@
-export declare const VERSION = "1.14.14";
-export declare const VERSION_DISPLAY = "v1.14.14";
+export declare const VERSION = "1.14.15";
+export declare const VERSION_DISPLAY = "v1.14.15";